CTRL+MouseWheel = Zoom if not in middle of draw operation.  The cursor is kept under the same doc pos after zoom in/out.  It may be a little off due to rounding error or if you scroll the wheel too quickly (and zoom operations are stacked I suppose?).
Backend:
* kpMainWindow
  - Add "centerUnderCursor" arg to zoomTo() - keeps cursor under same doc
       pos after zoom
  - Move {slotZoomIn,slotZoomOut,slotZoom}() functionality into
         {zoomIn,zoomOut,zoomAccordingToZoomAction}(bool centerUnderCursor)
* kpView
  - transform{ViewToDoc,DocToView}[XY](int) now accepts and returns doubles
  - Add transformViewToOtherView()
  - add wheelEvent()
* kpViewScrollableContainer
  - capture contentsWheelEvent() and pass to kpView::wheelEvent()
* kpTool
  - add wheelEvent()

[Update of my 000_ctrl_scroll=zoom.diff]


svn path=/branches/KDE/3.5/kdegraphics/kolourpaint/; revision=449622

+++ views/kpunzoomedthumbnailview.cpp
@@ -111,7 +111,7 @@ void kpUnzoomedThumbnailView::adjustToEnvironment ()
     int x;
     if (document ()->width () > width ())
     {
-        x = buddyView ()->transformViewToDocX (scrollViewContentsX);
+        x = (int) buddyView ()->transformViewToDocX (scrollViewContentsX);
         const int rightMostAllowedX = QMAX (0, document ()->width () - width ());
     #if DEBUG_KP_UNZOOMED_THUMBNAIL_VIEW
         kdDebug () << "\tdocX=" << x
@@ -134,7 +134,7 @@ void kpUnzoomedThumbnailView::adjustToEnvironment ()
     int y;
     if (document ()->height () > height ())
     {
-        y = buddyView ()->transformViewToDocY (scrollViewContentsY);
+        y = (int) buddyView ()->transformViewToDocY (scrollViewContentsY);
         const int bottomMostAllowedY = QMAX (0, document ()->height () - height ());
     #if DEBUG_KP_UNZOOMED_THUMBNAIL_VIEW
         kdDebug () << "\tdocY=" << y
